Move `EvaluateFieldContainer` to `field/_evaluate.py`
