+++ src/felupe/field/__init__.py
@@ -1,7 +1,8 @@
 from ._axi import FieldAxisymmetric
 from ._base import Field
-from ._container import EvaluateFieldContainer, FieldContainer
+from ._container import FieldContainer
 from ._dual import FieldDual
+from ._evaluate import EvaluateFieldContainer
 from ._fields import FieldsMixed
 from ._planestrain import FieldPlaneStrain
